test(pro): rename the native cache tests to the _native suffix
the _php.php and _python.py doublings existed to dodge the transpiled
twins that own the natural names (test_cache.php / test_cache.py are
generated from ts test.cache.ts) - but the suffix encodes provenance,
not language, and _native states that directly, stays correct as cases
accrue, and generalizes to every future hand-written sibling
