+++ php/pro/test/base/tests_init.php
@@ -9,7 +9,7 @@ include_once (__DIR__.'/test_order_book.php');
 include_once (__DIR__.'/test_cache.php');
 // php-only cache regressions - `test_cache.php` above is generated from
 // `ts/src/pro/test/base/test.cache.ts` and cannot host them
-include_once (__DIR__.'/test_cache_php.php');
+include_once (__DIR__.'/test_cache_native.php');
 // todo : include_once (__DIR__.'/test_close.php');
 include_once (__DIR__.'/test_client_retention.php');
